% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conRelPlot.r
\name{conRelPlot}
\alias{conRelPlot}
\title{Plotting conditional reliability}
\usage{
conRelPlot(
  model,
  theta_range = c(-4, 4),
  color = "red",
  title = "Conditional Reliability"
)
}
\arguments{
\item{model}{an object of class \code{SingleGroupClass} returned by the function \code{mirt()}.}

\item{theta_range}{range to be shown on the x-axis}

\item{color}{color of the line}

\item{title}{title for the plot (defaults to "Conditional Reliability")}
}
\value{
a ggplot
}
\description{
This function takes a fitted mirt-model and visualizes a conditional reliability curve. Heavily inspired by code from Phil Chalmers (author or 'mirt')
}
\examples{
library(mirt)
library(ggmirt)
data <- expand.table(LSAT7)
(mod <- mirt(data, 1))

conRelPlot(mod)
}
