% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemDistribution.r
\name{itemDist}
\alias{itemDist}
\title{Item difficulty distribution}
\usage{
itemDist(model, ...)
}
\arguments{
\item{model}{an object of class \code{SingleGroupClass} returned by the function \code{mirt()}.}

\item{...}{any argument passed to \code{geom_point()}}
}
\value{
a ggplot object.
}
\description{
This function requires a fitted mirt-model of class \code{SingleGroupClass} to visualize item difficulty distribution. Currently only works for unidimensional models.
}
\examples{
# Loading packages
library(mirt)
library(ggmirt)

# Getting data
data <- expand.table(LSAT7)

# Fitting a model 
(mod <- mirt(data, 1))

# Simple plot
itemDist(mod)

# Customized plot
itemDist(mod, size = 3, shape = 17, color = "blue")

}
