% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itempersonmap.r
\name{itemperson_map}
\alias{itemperson_map}
\title{Visualize item person map and scale properties based on Rasch model}
\usage{
itemperson_map(
  model,
  bin_color = "#0b66ff",
  no_bins = 45,
  title_left = "Item-Person-Map",
  title_right = "",
  scale_prop = TRUE
)
}
\arguments{
\item{model}{an object of class \code{SingleGroupClass} returned by the function \code{mirt()}.}

\item{bin_color}{color of the bins and points in the plot}

\item{no_bins}{number of bins to be displayed}

\item{title_left}{title of the item-person-map on the left.}

\item{title_right}{title of the scales properties on the left (defaults to none).}

\item{scale_prop}{logical value indicating whether scale properties should be plotted at all.}
}
\value{
a ggplot object.
}
\description{
This function takes a fitted mirt-model and visualizes and plots item-person-map (also known as Kernel-Density Plots or Wright maps) on the left, and add a scale characteristic curve, scale information curve, and a marginal reliability curve on the right.
}
\examples{
library(mirt)
data <- expand.table(LSAT7)
(mod1 <- mirt(data, 1))

itemperson_map(mod1, no_bins = 10)

}
