% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itempersonMap.r
\name{itempersonMap}
\alias{itempersonMap}
\title{Visualize item person map and scale properties based on Rasch model}
\usage{
itempersonMap(
  model,
  limits = c(-4, 4),
  title = "Item Person Map",
  margin = c(1, 0, -1.5, 0),
  density = FALSE,
  theme = theme_minimal(),
  ...
)
}
\arguments{
\item{model}{an object of class \code{SingleGroupClass} returned by the function \code{mirt()}.}

\item{limits}{range to be shown on the x-axis}

\item{title}{title for the plot (defaults to "Item-Person-Map")}

\item{margin}{margins around the top figure. Sometimes one might want to adjust this.}

\item{density}{logical value indicating whether a smoothed density curve or a standard histogram should be plotted.}

\item{theme}{any ggplot theme.}

\item{...}{any argument passed to \code{geom_point()}.}
}
\value{
a plot grid as returned by \code{cowplot::plot_grid()}
}
\description{
This function takes a fitted mirt-model and visualizes and plots item-person-map (also known as Kernel-Density Plots or Wright maps) on the left, and add a scale characteristic curve, scale information curve, and a marginal reliability curve on the right.
}
\examples{
library(mirt)
library(ggmirt)
data <- expand.table(LSAT7)
(mod <- mirt(data, 1))

itempersonMap(mod, 
              limits = c(-3, 3), 
              color = "red", 
              theme = theme_bw(), 
              shape = 17)

}
