% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personfitPlot.r
\name{personfitPlot}
\alias{personfitPlot}
\title{Plotting personfit estimates}
\usage{
personfitPlot(model, std = TRUE, title = "Person Infit and Outfit Statistics")
}
\arguments{
\item{model}{an object of class \code{SingleGroupClass} returned by the function \code{mirt()}.}

\item{std}{logical value indicating whether standardized or non-standardized infit or outfit estimates should be used (leads to different cut-off values).}

\item{title}{title for the plot (defaults to "Person Infit and Outfit Statistics")}
}
\value{
a ggplot
}
\description{
This function takes a fitted mirt-model and visualizes person infit and outfit estimates. The function builds on \code{mirt::itemfit()}. The basic idea is to visualize how many individuals in the sample do not show a response pattern that aligns with the suggested model. At best, the number of non-fitting response patterns is low (e.g., < 5\%).
}
\examples{
library(mirt)
library(ggmirt)
data <- expand.table(LSAT7)
(mod <- mirt(data, 1))

personfitPlot(mod)

}
\references{
\itemize{
\item Linacre JM. (2002). What do Infit and Outfit, Mean-square and Standardized mean? Rasch Measurement Transactions, 16(2), p.878. https://www.rasch.org/rmt/rmt162f.htm
}
}
