% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliPlot.r
\name{reliPlot}
\alias{reliPlot}
\title{Plotting conditional reliabilities across theta levels}
\usage{
reliPlot(
  model,
  color = "blue",
  theta_range = seq(-4, 4, 0.01),
  title = "Conditional reliability"
)
}
\arguments{
\item{model}{an object of class \code{SingleGroupClass} returned by the function \code{mirt()}.}

\item{color}{what color should the curve be?}

\item{theta_range}{range to be shown on the x-axis}

\item{title}{title for the plot (defaults to "Item Characteristic Curves")}
}
\value{
a ggplot
}
\description{
This function takes a fitted mirt-model plots reliability estimates across all levels of theta.
}
