% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_irt.r
\name{sim_irt}
\alias{sim_irt}
\title{Helper function to simulate IRT data}
\usage{
sim_irt(
  n.obs = 100,
  n.items = 10,
  discrimination = 0,
  seed = NULL,
  cut = "random"
)
}
\arguments{
\item{n.obs}{Number of observations that should be included in the data set}

\item{n.items}{Number of items that should be simulated}

\item{discrimination}{Standard deviation on the log scale}

\item{seed}{Seed for the random number generation process}

\item{cut}{Either "random" for a randomized transformation of the model probability matrix into the model 0-1 matrix or an integer value between 0 and 1.}
}
\value{
a tibble
}
\description{
Function to simulate data that can be used to fit IRT models.
}
\examples{
library(ggmirt)

sim_irt(n.obs = 200, n.items = 10)

}
