% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryPlot.r
\name{summaryPlot}
\alias{summaryPlot}
\title{A quick summary of IRT analyses}
\usage{
summaryPlot(model, theta_range = c(-4, 4), adj_factor = 0.05)
}
\arguments{
\item{model}{an object of class \code{SingleGroupClass} returned by the function \code{mirt()}.}

\item{theta_range}{range to be shown on the x-axis}

\item{adj_factor}{adjustment factor for properly overlaying information and standard error.}
}
\value{
a plot grid as returned by \code{cowplot::ggdraw()}
}
\description{
This function is essentially just a wrapper around several functions in this package and produces a summary of the most important aspects of an IRT model, including an item-person-map,test information curve, scale characteristic curve, and conditional reliability.
}
\examples{
library(mirt)
library(ggmirt)

# Simulate some data
data <- sim_irt(500, 10, seed = 123)

# Run IRT model with mirt
mod <- mirt(data, 1, itemtype = "2PL", verbose = FALSE)

summaryPlot(mod, theta_range = c(-4.5, 3.5), adj_factor = 1.5)

}
