% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testInfoCompare.r
\name{testInfoCompare}
\alias{testInfoCompare}
\title{Comparing test information curves of parallel tests}
\usage{
testInfoCompare(
  model1,
  model2,
  theta_range = c(-4, 4),
  title = "Parallel Tests",
  subtitle = "Test Information Curves"
)
}
\arguments{
\item{model1}{an object of class \code{SingleGroupClass} returned by the function \code{mirt()}.}

\item{model2}{an object of class \code{SingleGroupClass}}

\item{title}{title for the plot}

\item{subtitle}{subtitle for the plot}
}
\value{
a ggplot
}
\description{
This function takes two fitted mirt-model and visualizes test their test information curves on top of each other. This can be helpful for finding parallel tests.
}
\examples{
library(mirt)
library(ggmirt)
data <- expand.table(LSAT7)
(mod1 <- mirt(data, 1))
(mod2 <- mirt(data[,1:4], 1))

testInfoCompare(mod1, mod2)

}
