% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracePlot.r
\name{tracePlot}
\alias{tracePlot}
\title{Plotting item characteristics curves}
\usage{
tracePlot(
  model,
  data,
  theta_range = c(-4, 4),
  title = "Item Characteristics Curves",
  facet = TRUE,
  legend = FALSE
)
}
\arguments{
\item{model}{an object of class \code{SingleGroupClass} returned by the function \code{mirt()}.}

\item{data}{the data frame used to estimate the IRT model.}

\item{theta_range}{range to be shown on the x-axis}

\item{title}{title for the plot (defaults to "Item Characteristic Curves")}

\item{facet}{Should all items be shown in one plot, or each item received its individual facet?}
}
\value{
a ggplot
}
\description{
This function takes a fitted mirt-model and the underlying data and visualizes item characteristic curves.
}
\examples{
library(mirt)
library(ggmirt)
data <- expand.table(LSAT7)
(mod <- mirt(data, 1))

tracePlot(mod, data)
tracePlot(mod, data, theta_range = c(-5,5), facet = F, legend = T)

}
