% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personDistribution.r
\name{personDist}
\alias{personDist}
\title{Person parameter distribution}
\usage{
personDist(model, theta_range = c(-4, 4), density = FALSE, bins = 35)
}
\arguments{
\item{model}{an object of class \code{SingleGroupClass} returned by the function \code{mirt()}.}

\item{density}{logical value indicating whether a smoothed density curve or a standard histogram should be plotted.}

\item{bins}{number of bins to be plotted in the histogram}
}
\value{
a ggplot object.
}
\description{
This function requires a fitted mirt-model of class \code{SingleGroupClass} to visualize a person parameter distribution (theta levels in the studied population). The resulting ggplot can be further customized (e.g., with regard to theme, labels, etc.). It works with both uni- and multidimensional models.
}
\examples{
# Loading packages
library(mirt)
library(ggmirt)

# Getting data
data <- expand.table(LSAT7)

# Fitting a model 
(mod <- mirt(data, 1))

# Simple plot
personDist(mod)
personDist(mod, density = TRUE)

# Customized plot
personDist(mod, theta_range = c(-3, 3), bins = 10) +
  theme_classic()
}
