% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleCharPlot.r
\name{scaleCharPlot}
\alias{scaleCharPlot}
\title{Scale Characteristic Curve}
\usage{
scaleCharPlot(
  model,
  theta_range = c(-4, 4),
  color = "red",
  title = "Scale Characteristic Curve"
)
}
\arguments{
\item{model}{an object of class \code{SingleGroupClass} returned by the function \code{mirt()}.}

\item{theta_range}{range to be shown on the x-axis}

\item{color}{color of the line}

\item{title}{title for the plot (defaults to "Person Infit and Outfit Statistics")}
}
\value{
a ggplot
}
\description{
Once model-based theta score estimates are computed, it often is of interest to transform those estimates into the original scale metric. A scale characteristic function provides a means of transforming estimated theta scores to expected true scores in the original scale metric. This transformation back into the original scale metric provides a more familiar frame of reference for interpreting scores. This function provides a visualization for this transformation.
}
\examples{
library(mirt)
library(ggmirt)
data <- expand.table(LSAT7)
(mod <- mirt(data, 1))

scaleCharPlot(mod)

}
