% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryPlot.r
\name{summaryPlot}
\alias{summaryPlot}
\title{A quick summary of IRT analyses}
\usage{
summaryPlot(model, data)
}
\arguments{
\item{model}{an object of class \code{SingleGroupClass} returned by the function \code{mirt()}.}

\item{data}{the data frame used to estimate the IRT model.}
}
\value{
a plot grid as returned by \code{cowplot::ggdraw()}
}
\description{
This function is essentially just a wrapper around several functions in this package and produces a summary of the most important aspects of an IRT model, including an item-person-map, item characteristic curves, test information curve, and conditional reliability.
}
\examples{
library(mirt)
library(ggmirt)
data <- expand.table(LSAT7)
(mod <- mirt(data, 1))

summaryPlot(mod, data = data)

}
