% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracePlot.r
\name{tracePlot}
\alias{tracePlot}
\title{Plotting item characteristics curves}
\usage{
tracePlot(
  model,
  items = NULL,
  theta_range = c(-4, 4),
  title = "Item Characteristics Curves",
  n.answers = 5,
  facet = TRUE,
  legend = FALSE
)
}
\arguments{
\item{model}{an object of class \code{SingleGroupClass} returned by the function \code{mirt()}.}

\item{items}{numerical vector indicating which items to plot (currently does not yet work for graded response models).}

\item{theta_range}{range to be shown on the x-axis}

\item{title}{title for the plot (defaults to "Item Characteristic Curves")}

\item{n.answers}{In a graded response model, number of answer options (e.g., 5-point scale = 5)}

\item{facet}{Should all items be shown in one plot, or each item received its individual facet?}
}
\value{
a ggplot
}
\description{
This function takes a fitted mirt-model and the underlying data and visualizes item characteristic curves.
}
\examples{
library(mirt)
library(ggmirt)
data <- expand.table(LSAT7)
(mod <- mirt(data, 1))

tracePlot(mod)
tracePlot(mod, items = c(1,2,3), theta_range = c(-5,5), facet = F, legend = T)

}
